feat: drastically reduce traffic by collapsing chan chains
If we have a "string" result with this chain:

<-chan (<-chan (<-chan string))

The system will traverse this chain of values. Emitting a new channel
somewhere in the chain will cause the system to ignore the old channel
and traverse the new children for values.

Previously we transmitted a value node ID for each of these traversals,
this update collapses those into a single value node for the entire
chain. This simplifies the client a lot as well.